requirements.txt
# Web Framework
Flask>=3.0.0
Werkzeug>=3.0.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0

# AI Integration (optional - for AI features)
google-genai>=0.8.0

# Environment variables
python-dotenv>=1.0.0

# Data fetching
requests>=2.31.0
